requirements.txt
pytest
pylint
ipdb
jupyterlab
notebook
numpy
pandas

scikit-learn
ipdb
google-cloud-bigquery
google-cloud-storage
mlflow

### NLP ###
nltk

### STREAMLIT ###

streamlit

altair

plotly
plotly.express

#### API ###

uvicorn
fastapi
colorama
pytz
